+++ dir2cfg_template.py
@@ -4,13 +4,21 @@ import os
 import sys
 import yaml
 import argparse
-from refgenconf import CFG_GENOMES_KEY, CFG_SERVER_KEY, CFG_FOLDER_KEY, CFG_ARCHIVE_KEY, CFG_ASSET_PATH_KEY
+from refgenconf import (
+    CFG_GENOMES_KEY,
+    CFG_SERVER_KEY,
+    CFG_FOLDER_KEY,
+    CFG_ARCHIVE_KEY,
+    CFG_ASSET_PATH_KEY,
+)
 
 
 class _VersionInHelpParser(argparse.ArgumentParser):
     def format_help(self):
         """ Add version information to help text. """
-        return "version: {}\n".format(V) + super(_VersionInHelpParser, self).format_help()
+        return (
+            "version: {}\n".format(V) + super(_VersionInHelpParser, self).format_help()
+        )
 
 
 def build_dir2yaml_parser():
@@ -19,24 +27,25 @@ def build_dir2yaml_parser():
 
     :return argparse.ArgumentParser
     """
-    banner = "%(prog)s builds a refgenieserver config template for the directory structure." \
-             " Keep in mind that the produced config will require some adjustments."
+    banner = (
+        "%(prog)s builds a refgenieserver config template for the directory structure."
+        " Keep in mind that the produced config will require some adjustments."
+    )
     additional_description = "\nhttps://github.com/databio/refgenieserver"
 
-    parser = _VersionInHelpParser(
-        description=banner,
-        epilog=additional_description)
+    parser = _VersionInHelpParser(description=banner, epilog=additional_description)
 
     parser.add_argument(
-        "-V", "--version",
-        action="version",
-        version="%(prog)s {v}".format(v=V))
+        "-V", "--version", action="version", version="%(prog)s {v}".format(v=V)
+    )
     parser.add_argument(
-        "-p", "--path",
+        "-p",
+        "--path",
         dest="path",
         help="A path to the directory that the YAML should be build for. If not provided, current working directory "
-             "will be used ({})".format(os.getcwd()),
-        default=None)
+        "will be used ({})".format(os.getcwd()),
+        default=None,
+    )
     return parser
 
 
@@ -72,7 +81,7 @@ def main():
                     CFG_FOLDER_KEY: None,
                     CFG_SERVER_KEY: "http://www.refgenomes.databio.org",
                     CFG_ARCHIVE_KEY: None,
-                    CFG_GENOMES_KEY: None
+                    CFG_GENOMES_KEY: None,
                 }
                 rgc[CFG_GENOMES_KEY] = dir_as_dict(p)[os.path.basename(p)]
                 yaml.dump(rgc, f)