+++ refgenieserver/server_builder.py
@@ -5,8 +5,13 @@ from glob import glob
 from subprocess import run
 
 from refgenconf import RefGenConf
-from refgenconf.exceptions import RefgenconfError, ConfigNotCompliantError, \
-    GenomeConfigFormatError, MissingConfigDataError
+from refgenconf.exceptions import (
+    RefgenconfError,
+    ConfigNotCompliantError,
+    GenomeConfigFormatError,
+    MissingConfigDataError,
+)
+from refgenconf.helpers import swap_names_in_tree, replace_str_in_obj
 from attmap import PathExAttMap as PXAM
 from ubiquerg import checksum, size, is_command_callable, parse_registry_path
 
@@ -31,21 +36,36 @@ def archive(rgc, registry_paths, force, remove, cfg_path, genomes_desc):
     :param str cfg_path: config file path
     """
     if float(rgc[CFG_VERSION_KEY]) < float(REQ_CFG_VERSION):
-        raise ConfigNotCompliantError("You need to update the genome config to v{} in order to use the archiver. "
-                                      "The required version can be generated with refgenie >= {}".
-                                      format(REQ_CFG_VERSION, REFGENIE_BY_CFG[REQ_CFG_VERSION]))
+        raise ConfigNotCompliantError(
+            "You need to update the genome config to v{} in order to use the archiver. "
+            "The required version can be generated with refgenie >= {}".format(
+                REQ_CFG_VERSION, REFGENIE_BY_CFG[REQ_CFG_VERSION]
+            )
+        )
     if CFG_ARCHIVE_CONFIG_KEY in rgc:
         srp = rgc[CFG_ARCHIVE_CONFIG_KEY]
-        server_rgc_path = srp if os.path.isabs(srp) \
+        server_rgc_path = (
+            srp
+            if os.path.isabs(srp)
             else os.path.join(os.path.dirname(rgc.file_path), srp)
+        )
     else:
         try:
-            server_rgc_path = os.path.join(rgc[CFG_ARCHIVE_KEY], os.path.basename(cfg_path))
+            server_rgc_path = os.path.join(
+                rgc[CFG_ARCHIVE_KEY], os.path.basename(cfg_path)
+            )
         except KeyError:
-            raise GenomeConfigFormatError("The config '{}' is missing a {} entry. Can't determine the desired archive.".
-                                          format(cfg_path, " or ".join([CFG_ARCHIVE_KEY, CFG_ARCHIVE_KEY_OLD])))
+            raise GenomeConfigFormatError(
+                "The config '{}' is missing a {} entry. Can't determine the desired archive.".format(
+                    cfg_path, " or ".join([CFG_ARCHIVE_KEY, CFG_ARCHIVE_KEY_OLD])
+                )
+            )
     if os.path.isfile(server_rgc_path) and not os.access(server_rgc_path, os.W_OK):
-        raise OSError("The determined archive config path is not writable: {}".format(server_rgc_path))
+        raise OSError(
+            "The determined archive config path is not writable: {}".format(
+                server_rgc_path
+            )
+        )
     if force:
         _LOGGER.info("Build forced; file existence will be ignored")
     _LOGGER.debug("Registry_paths: {}".format(registry_paths))
@@ -56,16 +76,22 @@ def archive(rgc, registry_paths, force, remove, cfg_path, genomes_desc):
         rgc_server = RefGenConf(filepath=server_rgc_path)
         if remove:
             if not registry_paths:
-                _LOGGER.error("To remove archives you have to specify them. Use 'asset_registry_path' argument.")
+                _LOGGER.error(
+                    "To remove archives you have to specify them. Use 'asset_registry_path' argument."
+                )
                 exit(1)
             with rgc_server as r:
                 _remove_archive(r, registry_paths, CFG_ARCHIVE_KEY)
             exit(0)
     else:
         if remove:
-            _LOGGER.error("You can't remove archives since the genome_archive path does not exist yet.")
+            _LOGGER.error(
+                "You can't remove archives since the genome_archive path does not exist yet."
+            )
             exit(1)
-        _LOGGER.debug("'{}' file was not found and will be created".format(server_rgc_path))
+        _LOGGER.debug(
+            "'{}' file was not found and will be created".format(server_rgc_path)
+        )
         rgc_server = RefGenConf(filepath=rgc.file_path)
         rgc_server.make_writable(filepath=server_rgc_path)
         rgc_server.make_readonly()
@@ -81,28 +107,42 @@ def archive(rgc, registry_paths, force, remove, cfg_path, genomes_desc):
         exit(1)
     else:
         _LOGGER.debug("Genomes to be processed: {}".format(str(genomes)))
+    genomes = [rgc.get_genome_alias_digest(g) for g in genomes]
     if genomes_desc is not None:
         if os.path.exists(genomes_desc):
             import csv
-            _LOGGER.info("Found a genomes descriptions CSV file: {}".format(genomes_desc))
-            with open(genomes_desc, mode='r') as infile:
+
+            _LOGGER.info(
+                "Found a genomes descriptions CSV file: {}".format(genomes_desc)
+            )
+            with open(genomes_desc, mode="r") as infile:
                 reader = csv.reader(infile)
                 descs = {rows[0]: rows[1] for rows in reader}
         else:
-            _LOGGER.error("Genomes descriptions CSV file does not exist: {}".format(genomes_desc))
+            _LOGGER.error(
+                "Genomes descriptions CSV file does not exist: {}".format(genomes_desc)
+            )
             sys.exit(1)
     counter = 0
     for genome in genomes:
-        genome_dir = os.path.join(rgc[CFG_FOLDER_KEY], genome)
+        genome_dir = os.path.join(rgc.data_dir, genome)
         target_dir = os.path.join(rgc[CFG_ARCHIVE_KEY], genome)
+        alias_target_dir = os.path.join(
+            rgc[CFG_ARCHIVE_KEY], rgc.get_genome_alias(digest=genome, fallback=True)
+        )
         if not os.path.exists(target_dir):
             os.makedirs(target_dir, exist_ok=True)
-        genome_desc = rgc[CFG_GENOMES_KEY][genome].setdefault(CFG_GENOME_DESC_KEY, DESC_PLACEHOLDER) \
-            if genomes_desc is None or genome not in descs else descs[genome]
-        genome_checksum = rgc[CFG_GENOMES_KEY][genome].\
-            setdefault(CFG_CHECKSUM_KEY, CHECKSUM_PLACEHOLDER)
-        genome_attrs = {CFG_GENOME_DESC_KEY: genome_desc,
-                        CFG_CHECKSUM_KEY: genome_checksum}
+            # create legacy directory for archive
+            # TODO: remove in the future
+            os.makedirs(alias_target_dir, exist_ok=True)
+        genome_desc = (
+            rgc[CFG_GENOMES_KEY][genome].setdefault(
+                CFG_GENOME_DESC_KEY, DESC_PLACEHOLDER
+            )
+            if genomes_desc is None or genome not in descs
+            else descs[genome]
+        )
+        genome_attrs = {CFG_GENOME_DESC_KEY: genome_desc}
         with rgc_server as r:
             r[CFG_GENOMES_KEY].setdefault(genome, PXAM())
             r[CFG_GENOMES_KEY][genome].update(genome_attrs)
@@ -115,18 +155,29 @@ def archive(rgc, registry_paths, force, remove, cfg_path, genomes_desc):
         else:
             _LOGGER.debug("Assets to be processed: {}".format(str(assets)))
         for asset_name in assets if isinstance(assets, list) else [assets]:
-            asset_desc = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name]\
-                .setdefault(CFG_ASSET_DESC_KEY, DESC_PLACEHOLDER)
-            default_tag = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name]\
-                .setdefault(CFG_ASSET_DEFAULT_TAG_KEY, DEFAULT_TAG)
-            asset_attrs = {CFG_ASSET_DESC_KEY: asset_desc,
-                           CFG_ASSET_DEFAULT_TAG_KEY: default_tag}
-            _LOGGER.debug("Updating '{}/{}' asset attributes...".format(genome, asset_name))
+            asset_desc = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][
+                asset_name
+            ].setdefault(CFG_ASSET_DESC_KEY, DESC_PLACEHOLDER)
+            default_tag = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][
+                asset_name
+            ].setdefault(CFG_ASSET_DEFAULT_TAG_KEY, DEFAULT_TAG)
+            asset_attrs = {
+                CFG_ASSET_DESC_KEY: asset_desc,
+                CFG_ASSET_DEFAULT_TAG_KEY: default_tag,
+            }
+            _LOGGER.debug(
+                "Updating '{}/{}' asset attributes...".format(genome, asset_name)
+            )
             with rgc_server as r:
                 r.update_assets(genome, asset_name, asset_attrs)
 
             tag = tag_list[counter] if tag_list is not None else None
-            tags = tag or rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY].keys()
+            tags = (
+                tag
+                or rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ].keys()
+            )
             for tag_name in tags if isinstance(tags, list) else [tags]:
                 if not rgc.is_asset_complete(genome, asset_name, tag_name):
                     raise MissingConfigDataError(
@@ -135,68 +186,166 @@ def archive(rgc, registry_paths, force, remove, cfg_path, genomes_desc):
                         "refgenieserver archive requires all assets to be built"
                         " prior to archiving.".format(genome, asset_name, tag_name)
                     )
-                file_name = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY][tag_name][
-                    CFG_ASSET_PATH_KEY]
-                target_file = os.path.join(target_dir, "{}__{}".format(asset_name, tag_name) + ".tgz")
+                file_name = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ][tag_name][CFG_ASSET_PATH_KEY]
+                target_file = os.path.join(
+                    target_dir, "{}__{}".format(asset_name, tag_name) + ".tgz"
+                )
                 input_file = os.path.join(genome_dir, file_name, tag_name)
                 # these attributes have to be read from the original RefGenConf in case the archiver just increments
                 # an existing server RefGenConf
-                parents = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY][tag_name]. \
-                    setdefault(CFG_ASSET_PARENTS_KEY, [])
-                children = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY][tag_name]. \
-                    setdefault(CFG_ASSET_CHILDREN_KEY, [])
-                seek_keys = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY][tag_name]. \
-                    setdefault(CFG_SEEK_KEYS_KEY, {})
-                asset_digest = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][CFG_ASSET_TAGS_KEY][tag_name]. \
-                    setdefault(CFG_ASSET_CHECKSUM_KEY, None)
+                parents = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ][tag_name].setdefault(CFG_ASSET_PARENTS_KEY, [])
+                children = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ][tag_name].setdefault(CFG_ASSET_CHILDREN_KEY, [])
+                seek_keys = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ][tag_name].setdefault(CFG_SEEK_KEYS_KEY, {})
+                asset_digest = rgc[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                    CFG_ASSET_TAGS_KEY
+                ][tag_name].setdefault(CFG_ASSET_CHECKSUM_KEY, None)
                 if not os.path.exists(target_file) or force:
-                    _LOGGER.info("Creating archive '{}' from '{}' asset".format(target_file, input_file))
+                    _LOGGER.info(
+                        "Creating archive '{}' from '{}' asset".format(
+                            target_file, input_file
+                        )
+                    )
                     try:
-                        _check_tgz(input_file, target_file, asset_name)
+                        _check_tgz(input_file, target_file)
                         _copy_recipe(input_file, target_dir, asset_name, tag_name)
                         _copy_log(input_file, target_dir, asset_name, tag_name)
+                        # TODO: remove the rest of the try block in the future
+                        _check_tgz_legacy(
+                            input_file,
+                            target_file,
+                            asset_name,
+                            rgc.get_genome_alias_digest(alias=genome, fallback=True),
+                            rgc.get_genome_alias(digest=genome, fallback=True),
+                        )
+                        _copy_recipe(input_file, alias_target_dir, asset_name, tag_name)
+                        _copy_log(input_file, alias_target_dir, asset_name, tag_name)
                     except OSError as e:
                         _LOGGER.warning(e)
                         continue
                     else:
-                        _LOGGER.info("Updating '{}/{}:{}' tag attributes...".format(genome, asset_name, tag_name))
-                        tag_attrs = {CFG_ASSET_PATH_KEY: file_name,
-                                     CFG_SEEK_KEYS_KEY: seek_keys,
-                                     CFG_ARCHIVE_CHECKSUM_KEY: checksum(target_file),
-                                     CFG_ARCHIVE_SIZE_KEY: size(target_file),
-                                     CFG_ASSET_SIZE_KEY: size(input_file),
-                                     CFG_ASSET_PARENTS_KEY: parents,
-                                     CFG_ASSET_CHILDREN_KEY: children,
-                                     CFG_ASSET_CHECKSUM_KEY: asset_digest}
+                        _LOGGER.info(
+                            "Updating '{}/{}:{}' tag attributes...".format(
+                                genome, asset_name, tag_name
+                            )
+                        )
+                        tag_attrs = {
+                            CFG_ASSET_PATH_KEY: file_name,
+                            CFG_SEEK_KEYS_KEY: seek_keys,
+                            CFG_ARCHIVE_CHECKSUM_KEY: checksum(target_file),
+                            CFG_ARCHIVE_SIZE_KEY: size(target_file),
+                            CFG_ASSET_SIZE_KEY: size(input_file),
+                            CFG_ASSET_PARENTS_KEY: parents,
+                            CFG_ASSET_CHILDREN_KEY: children,
+                            CFG_ASSET_CHECKSUM_KEY: asset_digest,
+                        }
+                        # TODO: legacy checksum generation and tag dictionary
+                        #  update to be removed in the future
+                        legacy_digest = checksum(
+                            replace_str_in_obj(
+                                target_file,
+                                x=rgc.get_genome_alias_digest(
+                                    alias=genome, fallback=True
+                                ),
+                                y=rgc.get_genome_alias(digest=genome, fallback=True),
+                            )
+                        )
+                        tag_attrs.update(
+                            {CFG_LEGACY_ARCHIVE_CHECKSUM_KEY: legacy_digest}
+                        )
                         _LOGGER.debug("attr dict: {}".format(tag_attrs))
                         with rgc_server as r:
                             for parent in parents:
                                 # here we update any pre-existing parents' children attr with the newly added asset
-                                _LOGGER.debug("Updating {} children list with {}".
-                                              format(parent, "{}/{}:{}".format(genome, asset_name, tag_name)))
+                                _LOGGER.debug(
+                                    "Updating {} children list with {}".format(
+                                        parent,
+                                        "{}/{}:{}".format(genome, asset_name, tag_name),
+                                    )
+                                )
                                 rp = parse_registry_path(parent)
                                 parent_genome = rp["namespace"]
                                 parent_asset = rp["item"]
                                 parent_tag = rp["tag"]
                                 try:
-                                    r.seek(parent_genome, parent_asset, parent_tag, strict_exists=True)
+                                    r.seek(
+                                        parent_genome,
+                                        parent_asset,
+                                        parent_tag,
+                                        strict_exists=True,
+                                    )
                                 except RefgenconfError:
-                                    _LOGGER.warning("'{}/{}:{}'s parent '{}' does not exist, "
-                                                    "skipping relationship updates".
-                                                    format(genome, asset_name, tag_name, parent))
+                                    _LOGGER.warning(
+                                        "'{}/{}:{}'s parent '{}' does not exist, "
+                                        "skipping relationship updates".format(
+                                            genome, asset_name, tag_name, parent
+                                        )
+                                    )
                                     continue
-                                r.update_relatives_assets(parent_genome, parent_asset, parent_tag,
-                                                          ["{}/{}:{}".format(genome, asset_name, tag_name)],
-                                                          children=True)
+                                r.update_relatives_assets(
+                                    parent_genome,
+                                    parent_asset,
+                                    parent_tag,
+                                    ["{}/{}:{}".format(genome, asset_name, tag_name)],
+                                    children=True,
+                                )
                             r.update_tags(genome, asset_name, tag_name, tag_attrs)
                 else:
-                    _LOGGER.debug("'{}' exists".format(target_file))
+                    exists_msg = "'{}' exists.".format(target_file)
+                    try:
+                        rgc_server[CFG_GENOMES_KEY][genome][CFG_ASSETS_KEY][asset_name][
+                            CFG_ASSET_TAGS_KEY
+                        ][tag_name][CFG_ARCHIVE_CHECKSUM_KEY]
+                        _LOGGER.debug(exists_msg + " Skipping")
+                    except KeyError:
+                        _LOGGER.debug(exists_msg + " Calculating archive digest")
+                        tag_attrs = {CFG_ARCHIVE_CHECKSUM_KEY: checksum(target_file)}
+                        with rgc_server as r:
+                            r.update_tags(genome, asset_name, tag_name, tag_attrs)
+
         counter += 1
-    _LOGGER.info("Builder finished; server config file saved to {}".format(rgc_server.file_path))
+    _LOGGER.info(
+        "Builder finished; server config file saved to {}".format(rgc_server.file_path)
+    )
 
 
-def _check_tgz(path, output, asset_name):
+def _check_tgz(path, output):
     """
+    Check if file exists and tar it.
+    If gzipping is requested, the pigz software is used if available.
+
+    :param str path: path to the file to be tarred
+    :param str output: path to the result file
+    :raise OSError: if the file/directory meant to be archived does not exist
+    """
+    pth, tag_name = os.path.split(path)
+    if os.path.exists(path):
+        # tar gzip the asset, exclude _refgenie_build dir, it may change digests
+        cmd = "tar --exclude '_refgenie_build' -C {p} "
+        cmd += (
+            "-cvf - {tn} | pigz > {o}"
+            if is_command_callable("pigz")
+            else "-cvzf {o} {tn}"
+        )
+        command = cmd.format(p=pth, o=output, tn=tag_name)
+        _LOGGER.info("command: {}".format(command))
+        run(command, shell=True)
+    else:
+        raise OSError("Entity '{}' does not exist".format(path))
+
+
+def _check_tgz_legacy(path, output, asset_name, genome_name, alias):
+    """
+    NOTE: This is a duplication of the _check_tgz function, kept separate as in
+    the future this step will be simply removed.
+
     Check if file exists and tar it.
     If gzipping is requested, the availability of pigz software is checked and used.
 
@@ -205,23 +354,30 @@ def _check_tgz(path, output, asset_name):
     :param str asset_name: name of the asset
     :raise OSError: if the file/directory meant to be archived does not exist
     """
-    # since the genome directory structure changed (added tag layer) in refgenie >= 0.7.0 we need to perform some
-    # extra file manipulation before archiving to make the produced archive compatible with new and old versions
-    # of refgenie CLI. The difference is that refgenie < 0.7.0 requires the asset to be archived with the asset-named
-    # enclosing dir, but with no tag-named directory as this concept did not exist back then.
-    if os.path.exists(path):
-        # copy the asset files to asset-named dir, excluding _refgenie_build dir, which may change digests
-        cmd = "rsync -rvL --exclude '_refgenie_build' {p}/ {p}/{an}/; cd {p}; "
-        # tar gzip the new dir
-        cmd += "tar -cvf - {an} | pigz > {o}; " if is_command_callable("pigz") \
-            else "tar -cvzf {o} {an}; "
-        # remove the new dir
-        cmd += "rm -r {p}/{an}"
-        command = cmd.format(p=path, o=output, an=asset_name)
-        _LOGGER.debug("command: {}".format(command))
-        run(command, shell=True)
-    else:
-        raise OSError("Entity '{}' does not exist".format(path))
+    # TODO: remove in the future
+    if isinstance(alias, str):
+        alias = [alias]
+    for a in alias:
+        if os.path.exists(path):
+            aliased_output = replace_str_in_obj(output, x=genome_name, y=a)
+            cmd = "rsync -rvL --exclude '_refgenie_build' {p}/ {p}/{an}/; "
+            command = cmd.format(p=path, o=output, an=asset_name)
+            _LOGGER.debug("command: {}".format(command))
+            run(command, shell=True)
+            swap_names_in_tree(os.path.join(path, asset_name), a, genome_name)
+            # tar gzip the new dir
+            cmd = (
+                "cd {p}; tar -cvf - {an} | pigz > {oa}; "
+                if is_command_callable("pigz")
+                else "tar -cvzf {oa} {an}; "
+            )
+            # remove the new dir
+            cmd += "rm -r {p}/{an}"
+            command = cmd.format(p=path, oa=aliased_output, an=asset_name)
+            _LOGGER.debug("command: {}".format(command))
+            run(command, shell=True)
+        else:
+            raise OSError("Entity '{}' does not exist".format(path))
 
 
 def _copy_log(input_dir, target_dir, asset_name, tag_name):
@@ -233,8 +389,13 @@ def _copy_log(input_dir, target_dir, asset_name, tag_name):
     """
     log_path = "{}/{}/{}".format(input_dir, BUILD_STATS_DIR, ORI_LOG_NAME)
     if log_path and os.path.exists(log_path):
-        run("cp " + log_path + " " +
-            os.path.join(target_dir, TEMPLATE_LOG.format(asset_name, tag_name)), shell=True)
+        run(
+            "cp "
+            + log_path
+            + " "
+            + os.path.join(target_dir, TEMPLATE_LOG.format(asset_name, tag_name)),
+            shell=True,
+        )
         _LOGGER.debug("Log copied to: {}".format(target_dir))
     else:
         _LOGGER.warning("Log not found: {}".format(log_path))
@@ -249,7 +410,9 @@ def _copy_recipe(input_dir, target_dir, asset_name, tag_name):
     :param str asset_name: asset name
     :param str tag_name: tag name
     """
-    recipe_path = "{}/{}/{}".format(input_dir, BUILD_STATS_DIR, TEMPLATE_RECIPE_JSON.format(asset_name, tag_name))
+    recipe_path = "{}/{}/{}".format(
+        input_dir, BUILD_STATS_DIR, TEMPLATE_RECIPE_JSON.format(asset_name, tag_name)
+    )
     if recipe_path and os.path.exists(recipe_path):
         run("cp " + recipe_path + " " + target_dir, shell=True)
         _LOGGER.debug("Recipe copied to: {}".format(target_dir))
@@ -268,17 +431,36 @@ def _remove_archive(rgc, registry_paths, cfg_archive_folder_key=CFG_ARCHIVE_KEY)
     """
     ret = []
     for registry_path in _correct_registry_paths(registry_paths):
-        genome, asset, tag = registry_path["namespace"], registry_path["item"], registry_path["tag"]
+        genome, asset, tag = (
+            registry_path["namespace"],
+            registry_path["item"],
+            registry_path["tag"],
+        )
         try:
             if asset is None:
-                [rgc.cfg_remove_assets(genome, x, None) for x in rgc.list_assets_by_genome(genome)]
+                [
+                    rgc.cfg_remove_assets(genome, x, None)
+                    for x in rgc.list_assets_by_genome(genome)
+                ]
             else:
                 rgc.cfg_remove_assets(genome, asset, tag)
-            _LOGGER.info("{}/{}{} removed".format(genome, asset, ":" + tag if tag else ""))
+            _LOGGER.info(
+                "{}/{}{} removed".format(genome, asset, ":" + tag if tag else "")
+            )
         except KeyError:
-            _LOGGER.warning("{}/{}{} not found and not removed.".format(genome, asset, ":" + tag if tag else ""))
+            _LOGGER.warning(
+                "{}/{}{} not found and not removed.".format(
+                    genome, asset, ":" + tag if tag else ""
+                )
+            )
             continue
-        ret.append(os.path.join(rgc[cfg_archive_folder_key], genome, "{}__{}".format(asset or "*", tag or "*") + ".tgz"))
+        ret.append(
+            os.path.join(
+                rgc[cfg_archive_folder_key],
+                genome,
+                "{}__{}".format(asset or "*", tag or "*") + ".tgz",
+            )
+        )
         for p in ret:
             archives = glob(p)
             for path in archives:
@@ -302,6 +484,7 @@ def _correct_registry_paths(registry_paths):
     :param list[dict] registry_paths: output of parse_registry_path
     :return list[dict]: corrected registry paths
     """
+
     def _swap(rp):
         """
         Swaps dict values of 'namespace' with 'item' keys
@@ -325,6 +508,3 @@ def _get_paths_element(registry_paths, element):
     :return list[str]: extracted elements
     """
     return [x[element] for x in _correct_registry_paths(registry_paths)]
-
-
-